refactor: move create_random_state_diff to papyrus_common
