+++ crates/papyrus_common/Cargo.toml
@@ -17,6 +17,8 @@ sha3.workspace = true
 starknet_api.workspace = true
 starknet-crypto.workspace = true
 thiserror.workspace = true
+rand.workspace = true
+indexmap.workspace = true
 
 [dev-dependencies]
 assert_matches.workspace = true
