+++ crates/papyrus_common/src/state.rs
@@ -1,7 +1,9 @@
+use indexmap::indexmap;
+use rand::RngCore;
 use serde::{Deserialize, Serialize};
-use starknet_api::core::{ClassHash, CompiledClassHash, ContractAddress};
+use starknet_api::core::{ClassHash, CompiledClassHash, ContractAddress, Nonce};
 use starknet_api::hash::StarkFelt;
-use starknet_api::state::StorageKey;
+use starknet_api::state::{StorageKey, ThinStateDiff};
 
 /// A storage entry in a contract.
 #[derive(Debug, Default, Clone, Eq, PartialEq, Hash, Deserialize, Serialize, PartialOrd, Ord)]
@@ -28,4 +30,32 @@ pub struct DeclaredClassHashEntry {
 pub struct ReplacedClass {
     pub address: ContractAddress,
     pub class_hash: ClassHash,
+}
+
+pub fn create_random_state_diff(rng: &mut impl RngCore) -> ThinStateDiff {
+    let contract0 = ContractAddress::from(rng.next_u64());
+    let contract1 = ContractAddress::from(rng.next_u64());
+    let contract2 = ContractAddress::from(rng.next_u64());
+    let class_hash = ClassHash(rng.next_u64().into());
+    let compiled_class_hash = CompiledClassHash(rng.next_u64().into());
+    let deprecated_class_hash = ClassHash(rng.next_u64().into());
+    ThinStateDiff {
+        deployed_contracts: indexmap! {
+            contract0 => class_hash, contract1 => class_hash, contract2 => deprecated_class_hash
+        },
+        storage_diffs: indexmap! {
+            contract0 => indexmap! {
+                1u64.into() => StarkFelt::ONE, 2u64.into() => StarkFelt::TWO
+            },
+            contract1 => indexmap! {
+                3u64.into() => StarkFelt::TWO, 4u64.into() => StarkFelt::ONE
+            },
+        },
+        declared_classes: indexmap! { class_hash => compiled_class_hash },
+        deprecated_declared_classes: vec![deprecated_class_hash],
+        nonces: indexmap! {
+            contract0 => Nonce(StarkFelt::ONE), contract2 => Nonce(StarkFelt::TWO)
+        },
+        replaced_classes: Default::default(),
+    }
 }
